Moved semi-random function to new xdem/misc.py
