+++ tests/test_coreg.py
@@ -22,7 +22,7 @@ import pytransform3d.transformations
 
 with warnings.catch_warnings():
     warnings.simplefilter("ignore")
-    from xdem import coreg, examples, spatial_tools
+    from xdem import coreg, examples, spatial_tools, misc
 
 
 def load_examples() -> tuple[gu.georaster.Raster, gu.georaster.Raster, gu.geovector.Vector]:
@@ -612,28 +612,6 @@ def test_apply_matrix():
     print(np.nanmedian(diff), spatial_tools.nmad(diff))
 
 
-def generate_random_field(shape: tuple[int, int], corr_size: int) -> np.ndarray:
-    """
-    Generate a semi-random gaussian field (to simulate a DEM or DEM error)
-
-    :param shape: The output shape of the field.
-    :param corr_size: The correlation size of the field.
-
-    :returns: A numpy array of semi-random values from 0 to 1
-    """
-    field = cv2.resize(
-        cv2.GaussianBlur(
-            np.repeat(np.repeat(
-                np.random.randint(0, 255, (shape[0]//corr_size,
-                                           shape[1]//corr_size), dtype='uint8'),
-                corr_size, axis=0), corr_size, axis=1),
-            ksize=(2*corr_size + 1, 2*corr_size + 1),
-            sigmaX=corr_size) / 255,
-        dsize=(shape[1], shape[0])
-    )
-    return field
-
-
 def test_warp_dem():
     """Test that the warp_dem function works expectedly."""
     warnings.simplefilter("error")
@@ -708,7 +686,7 @@ def test_warp_dem():
     # Generate a semi-random DEM
     transform = rio.transform.from_origin(0, 500, 1, 1)
     shape = (500, 550)
-    dem = generate_random_field(shape, 100) * 200 + generate_random_field(shape, 10) * 50
+    dem = misc.generate_random_field(shape, 100) * 200 + misc.generate_random_field(shape, 10) * 50
 
     # Warp the DEM using the source-destination coordinates.
     transformed_dem = coreg.warp_dem(
